Add more authenticated channels, some adjustmets

+++ csharp/src/Org.OpenAPITools/com.blockchain.exchange.rest.model/Price.cs
@@ -1,9 +1,9 @@
 ﻿namespace Org.OpenAPITools.Client;
 
-public sealed record Price(long Timestamp, double Open, double High, double Low,
+public sealed record Price(double Timestamp, double Open, double High, double Low,
 	double Close, double Volume)
 {
-	public long Timestamp { get; set; } = Timestamp;
+	public double Timestamp { get; set; } = Timestamp;
 	public double Open { get; set; } = Open;
 	public double High { get; set; } = High;
 	public double Low { get; set; } = Low;
